+++ timeline/plugins.py
@@ -11,7 +11,7 @@ from projectroles.utils import get_display_name
 
 from timeline.api import TimelineAPI
 from timeline.models import ProjectEvent
-from timeline.urls import urlpatterns
+from timeline.urls import urls_ui_project, urls_ui_site, urls_ui_admin
 
 
 class ProjectAppPlugin(ProjectAppPluginPoint):
@@ -26,7 +26,7 @@ class ProjectAppPlugin(ProjectAppPluginPoint):
     title = 'Timeline'
 
     #: App URLs (will be included in settings by djangoplugins)
-    urls = urlpatterns
+    urls = urls_ui_project
 
     # Properties defined in ProjectAppPluginPoint -----------------------
 
@@ -151,7 +151,7 @@ class SiteAppPlugin(SiteAppPluginPoint):
     title = 'Site-Wide Events'
 
     #: App URLs (will be included in settings by djangoplugins)
-    urls = urlpatterns
+    urls = urls_ui_site
 
     #: Iconify icon
     icon = 'mdi:clock-time-eight-outline'
@@ -176,7 +176,7 @@ class AdminSiteAppPlugin(SiteAppPluginPoint):
     title = 'All Timeline Events'
 
     #: App URLs (will be included in settings by djangoplugins)
-    urls = urlpatterns
+    urls = urls_ui_admin
 
     #: Iconify icon
     # icon = 'mdi:clock-star-four-points-outline'
@@ -186,7 +186,7 @@ class AdminSiteAppPlugin(SiteAppPluginPoint):
     description = 'Admin view for all timeline events on the site'
 
     #: Entry point URL ID
-    entry_point_url_id = 'timeline:timeline_site_admin'
+    entry_point_url_id = 'timeline:list_admin'
 
     #: Required permission for displaying the app
     app_permission = 'timeline.view_site_admin'